crew/pyproject.toml
[project]
name = "crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai-tools>=0.48.0",
    "streamlit>=1.46.0",
]
